structure changes + redid tag type
i think tags would work better as a map of sets. it'd be more organized this way. i also redid the file structure to how (i think) the project description wants it. App.java and all the class files are in the models folder, and the controller and fxml files are in the view folder right now.